pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "dash>=2.9.3",
    "email-validator>=2.2.0",
    "flask-socketio>=5.5.1",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "python-socketio>=5.13.0",
    "dash-bootstrap-components>=1.7.1",
    "plotly>=6.1.2",
    "numpy>=2.2.6",
    "pyyaml>=6.0.2",
    "werkzeug>=3.1.3",
    "socketio>=0.2.1",
]
